[DF-520] Changed uint8 to uint32 due to uint8 not added to VariantType
Former-commit-id: 1659718087e3f3a2e054f75acb0e7f60629faf45

+++ Sources/Internal/Render/Highlevel/RenderBatch.h
@@ -120,7 +120,7 @@ protected:
     uint32 indexCount;
     
 //    ePrimitiveType type; //TODO: waiting for enums at introspection
-    uint8 type;
+    uint32 type;
     
     RenderLayer * ownerLayer;
     uint32 removeIndex;
